Move display to auto shuffleboard
